Rename time to chronology
