+++ meson.build
@@ -33,7 +33,7 @@ types_inc = include_directories('types/include')
 sources = [
     'types/src/ports.c',
     'types/src/interpolation.c',
-    'types/src/time.c',
+    'types/src/chronology.c',
     'types/src/test_factories.c',
     'types/src/atoms.c'
 ]
