+++ types/include/atoms.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "time.h"
+#include "chronology.h"
 #include "macros.h"
 
 typedef union {
