+++ types/include/interpolation.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "time.h"
+#include "chronology.h"
 #include "atoms.h"
 #include "error.h"
 #include "time_series.h"
