+++ types/include/time_series.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "time.h"
+#include "chronology.h"
 #include "atoms.h"
 #include "macros.h"
